# openapi.yaml
openapi: 3.0.3
info:
  title: Todo OpenApi
  description: |
    CRUD for todo using Cbor. More at [https://w4.ssegning.com](https://w4.ssegning.com)
  termsOfService: https://w4.ssegning.com/res/tos
  contact:
    email: hello@vymalo.com
    name: Vymalo Todo Team
  license:
    name: MIT
  version: 0.1.0
...
servers: [ ]
...
paths:
...
  /api/todos/{id}:
    get:
      tags:
        - todos
      summary: Get todo
      operationId: getTodo
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: 'successful operation'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
        400:
          description: 'wrong content provided'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
        500:
          description: 'Internal server'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
    delete:
      tags:
        - todos
      summary: Delete todo
      operationId: deleteTodo
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        202:
          description: 'successful operation'
        400:
          description: 'wrong content provided'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
        500:
          description: 'Internal server'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
    put:
      tags:
        - todos
      summary: Update todo
      operationId: updateTodo
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/cbor:
            schema:
              type: string
              format: binary
        required: true
      responses:
        200:
          description: 'successful operation'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
        400:
          description: 'wrong content provided'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
        500:
          description: 'Internal server'
          content:
            application/cbor:
              schema:
                type: string
                format: binary
components:
  parameters:
...
    id:
      name: id
      in: path
      required: true
      description: The slug name or id of the todo
      schema:
        $ref: '#/components/schemas/Cuid'
  schemas:
...
    Cuid:
      type: string
      format: cuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT